Fixed Bugs
Bug:
The Writing of Workers to File Would Result in an Infinite Loop (Writing of the "First" Worker to File Infinite Times).

Reason:
Iterators Used to Traverse the Unordered Sets "admins" and "deliveryPeople" Were Not Being Incremented on Base::writeWorkersFile.

-------------------------------

Bug:
Wrong Output when Writing to the Workers Text File.

Reason:
Extra "endl" After the Day, Month and Year of the Birthday.