Fixed Bugs
Bug:
Wrong Output when Writing to the Workers Text File.

Reason:
Missing "endl" After the Year of the Birthday.

-------------------------------

Bug:
Wrong Output when Writing to the Workers Text File.

Reason:
In the Case of Only One Element in the "(*delivIt)->get_history()" and "(*delivIt)" Not Corresponding to the First Element in the Data Structure, it Would Output " : [number]" Instead of the Intended "[number]" Which Would Cause a Crash When Trying To Read From the "Defective" File.